% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget_calc.R
\name{budget_calc}
\alias{budget_calc}
\title{Calculate Total Budget}
\usage{
budget_calc(
  n_primary,
  n_secondary,
  n_followup_primary = 1,
  n_followup_secondary = 1
)
}
\arguments{
\item{n_primary}{Number of participants in the study per group.}

\item{n_secondary}{Number of participants receiving an MRI per group.}

\item{n_followup_primary}{Number of follow up visits after the enrollment visit where CDR-SB is tested.}

\item{n_followup_secondary}{Number of follow up visits after the enrollment visit where an MRI is scheduled.}
}
\value{
Total budget for the specified study design.
}
\description{
Calculate Total Budget
}
