% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw4-package.R
\docType{package}
\name{hw4-package}
\alias{hw4}
\alias{hw4-package}
\title{hw4: Study Planning Analysis for Alzheimers Treatment}
\description{
Allows users to vary study parameters and see their impact on the power and budget of an Alzheimers treatment trial.
}
\author{
\strong{Maintainer}: Sara Hunsberger \email{shunsbe2@jh.edu}

}
\keyword{internal}
