% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n.R
\name{plot_n}
\alias{plot_n}
\title{Make a heatmap showing the number of participants needed for different delta and sd values}
\usage{
plot_n(
  power = 0.9,
  delta_start = 0.1,
  delta_end = 1,
  sd_start = 0.5,
  sd_end = 1.5,
  length_vec = 20
)
}
\arguments{
\item{power}{The power for the test you want to achieve}

\item{delta_start}{Starting value of the values of deltas plotted.}

\item{delta_end}{Ending value of the values of deltas plotted.}

\item{sd_start}{Starting value of the values of standard deviations plotted.}

\item{sd_end}{Ending value of the values of standard deviations plotted.}

\item{length_vec}{Number of values of delta and sd plotted.}
}
\value{
Heatmap of the n value needed to get the power value specified at different values of delta and sd.
}
\description{
Make a heatmap showing the number of participants needed for different delta and sd values
}
