% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calc.R
\name{sample_size_calc}
\alias{sample_size_calc}
\title{Calculate Sample Size}
\usage{
sample_size_calc(
  delta = 0.5,
  sd = 1.08,
  sig.level = 0.05,
  power = 0.9,
  bonferonni = FALSE,
  alternative = "two.sided"
)
}
\arguments{
\item{delta}{Estimate of the difference in change in outcome annually from treatment to placebo groups. Default is set to 0.5 clinically meaningful change in CDR-SB score.}

\item{sd}{Estimate of the standard deviation in outcome annually for a person with mild cognitive impairment. Default is set to 1.08 (a sd of change in CDR-SB score annually).}

\item{sig.level}{Significance level of the test.}

\item{power}{Power of test (1 minus Type II error probability)}

\item{bonferonni}{Boolean specifying if the user wants to include a bonferonni correction for a secondary outcome.}

\item{alternative}{One- or two-sided test.}
}
\value{
Number of participants needed in each group.
}
\description{
Calculate Sample Size
}
